+++ src/commands/helps/init.rs
@@ -25,5 +25,6 @@ Options:
     <path>                directory to initialize
     --name <name>         set the resulting ip's name
     --lib <lib>           set the resulting ip's library
+    --uuid                print a new uuid and exit
 
 Use 'orbit help init' to read more about the command."#;